refactor: organize registry source by category
