+++ packages/cli/src/helpers/turbo-affected.test.ts
@@ -71,8 +71,8 @@ test.each([
 	["packages/cli/src/index.ts", true],
 	["packages/registry/registry.ts", true],
 	["packages/registry/metadata.ts", true],
-	["packages/registry/src/word-count/tool.ts", true],
-	["packages/registry/gateways/vercel.ts", true],
+	["packages/registry/src/tools/word-count/tool.ts", true],
+	["packages/registry/src/gateways/vercel/gateway.ts", true],
 	["packages/cli/scripts/test-scaffold.sh", true],
 	["apps/chat/app/page.tsx", true],
 	["apps/chat/package.json", true],
