+++ packages/registry/src/gateways/litellm/gateway.ts
