+++ packages/registry/src/gateways/openai-compatible/gateway.ts
