+++ packages/registry/src/gateways/openai/gateway.ts
