+++ packages/registry/src/gateways/openrouter/gateway.ts
