+++ packages/registry/src/tools/get-weather/renderer.tsx
