+++ packages/registry/src/tools/get-weather/tool.ts
