+++ packages/registry/src/tools/word-count/renderer.tsx
