+++ packages/registry/src/tools/word-count/tool.test.ts
