+++ packages/registry/src/tools/word-count/tool.ts
